NX Monorepo created, migration from Development branch in progress
Created an NX workspace to simplify project management and also because, due to some misconfiguration, components just don't load when included in a page. Currently the only way to load components is to define them in a module which then gets loaded by the router, meaning they can't be reused in other pages. 
Thus far only the Home and Login pages have been migrated from development to the monorepo.